requirements_signvlm.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
peft>=0.4.0
tokenizers>=0.13.0
accelerate>=0.20.0
datasets>=2.12.0
wandb>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
opencv-python>=4.5.0
mediapipe>=0.10.0
Pillow>=8.3.0
tqdm>=4.62.0
nltk>=3.7.0
rouge-score>=0.1.2
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0 